+++ src/MicroTeX/lib/atom/font_family_atom.cpp
@@ -1,4 +1,4 @@
-#include "font_family_atom.h"
+#include "atom/font_family_atom.h"
 
 #include "atom/atom_font.h"
 #include "box/box_single.h"