+++ src/MicroTeX/lib/macro/macro_env.h
@@ -5,12 +5,14 @@
 #include <functional>
 #include <vector>
 
+#include "atom/atom_basic.h"
 #include "atom/atom_matrix.h"
 #include "core/formula.h"
 #include "core/parser.h"
 #include "macro/macro.h"
 #include "macro/macro_decl.h"
 #include "utils/exceptions.h"
+#include "utils/string_utils.h"
 
 namespace microtex {